fixed itoa base to take larger numbers

+++ libft/ft_itoa_base.c
@@ -6,18 +6,18 @@
 /*   By: sjones <sjones@student.42.us.org>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2017/11/27 13:29:58 by sjones            #+#    #+#             */
-/*   Updated: 2017/11/27 13:30:33 by sjones           ###   ########.fr       */
+/*   Updated: 2017/11/27 13:54:13 by sjones           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-char*ft_itoa_base(int value, int base)
+char	*ft_itoa_base(long long int value, int base)
 {
-	long int	n;
-	char		*ret;
-	int			len;
-	int			i;
+	unsigned long long int	n;
+	char					*ret;
+	int						len;
+	int						i;
 
 	n = ABS(value);
 	len = 1;
